Added a biased FM detector This detector adds a positive or negative offset to the data slicer threshold depending on the calculated reference bitstream. This removes many bit errors, especially in cases where the correlation ratio is near the threshold. It luckily does not seem to turn pure noise into "valid" data!
This detector type will likely be used in a final system once a good initial signal acquisition has been done.
However, obviously I still plan to continue work on the raw detector performance, as I still feel there is more to be done.
It's reasonable to assume that a timing receiver includes a battery backed RTC, so knowing the time to within a minute or two is expected.