Add a new aggressive autobias detector The FM Biased MeanVariance detector works by doing two passes over the minute data. After the first pass, it finds the mean value of the ratios and sets this as the sampler threshold. It also computes the variance of the data, and uses this as the autobias offset. This detector is better for handling the odd cases where the average ratio ends up being e.g. centered at 0.5 instead of 1 like it normally is. Also added the basic Mean type, which only changes the threshold based on the mean of the ratios.
Changed the value we insert for max_zero ratio of 0 to be the threshold+1. With the old value of 999 the variance would go crazy, and the new detector would "correctly" decode pure noise.

This function might be of use in a real system if some limits are placed (e.g. only enabling it when the SNR is within a reasonable range). It would be implemented as a FIR filter with the ratios as input.

Also added a helper for checking if a demodulator is basically an FM type with extensions to simplify e.g. the correlator function checks.