Add high pass filtering to correlator output This was added because of test signal 24, which is nearly perfect. Unfortunately it also completely messed up the performance of the standard FM demodulator!
The cause was that there was significant variation in the DC output of the correlator through the pass, especially near the minute sync.
The ratio function works best with DC-free signals.
The good news is that this change completely fixed the good-SNR performance, the bad news is a few noisy recordings did get slightly noisier.
However both biased and biased mean detectors now perform even better!
Amazingly, recording 14 can now be decoded with no errors!