Add some alternate correlation methods Our current (and still best) detector is more appropriate called a square sum of absolute difference detector. The "correct" and most sensitive detector is the multiply-accumulate detector.
The problem with the MAC correlator is that it ignores any template coefficients with a weight of zero. In our case we actually do care about zeros, and the SSAD detector does weight zeros equally.

A hybrid was also tested which uses the SSAD method for weights close to zero, and MAC for the non-zero weights. This is then a SSAD MAC detector :)