Simplified minute start detector It even works better!
This detector runs a correlator in the form of squaring the envelope signal with a range equal to the length of the old correlator.
The result is offset by the sum (though this is not required in principle).

The detector then weights the result by subtracting the  value found at -50 (divided by 5). This makes it more "peaky".
An offset is applied to the output sample number to align it to the old reference mark.